+++ src/RoomDetail/components/RoomMembers.tsx
@@ -1,18 +1,14 @@
 import { useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import { useLocalStorage, useRouteData } from '@/core/hooks';
-import notificationAPI from '@/core/api/functions/notificationAPI';
 import { RankMember } from '@/core/types/Member';
-import { Avatar } from '@/shared/Avatar';
-import { Icon } from '@/shared/Icon';
 import { useBottomSheet } from '@/shared/BottomSheet';
-import { Toast } from '@/shared/Toast';
 import ReportBottomSheet from './ReportBottomSheet';
+import RoomMember from './RoomMember';
 
 interface RoomMembers {
   members: RankMember[];
   reportStatus: boolean;
-  changeReportStatus: (value: boolean) => void;
+  changeReportStatus: () => void;
 }
 
 const RoomMembers = ({
@@ -41,27 +37,6 @@ const RoomMembers = ({
     }
   });
   const [checked, setChecked] = useState('');
-  const [myUserId] = useLocalStorage('MEMBER_ID', null);
-
-  const {
-    params: { roomId }
-  } = useRouteData();
-
-  const handlePokeButtonClick = async (memberId: number, nickname: string) => {
-    const { status } = await notificationAPI.getMemberPoke(
-      roomId || '',
-      memberId
-    );
-
-    if (status === 200) {
-      Toast.show({
-        status: 'confirm',
-        message: `${nickname}을 콕! 찔렀어요`,
-        icon: true,
-        subText: '콕콕'
-      });
-    }
-  };
 
   const changeCheckedInput = (state: string) => {
     setChecked(state);
@@ -80,69 +55,14 @@ const RoomMembers = ({
   return (
     <>
       <div className="mt-[2.87rem]">
-        {members.map(
-          ({
-            memberId,
-            nickname,
-            profileImage,
-            contributionPoint,
-            rank,
-            isNotificationSent
-          }) => (
-            <div
-              key={memberId}
-              className="mb-[1.19rem] flex items-center justify-between"
-            >
-              <Avatar
-                imgUrl={profileImage}
-                userId={memberId}
-                nickname={nickname}
-                contribution={contributionPoint}
-              />
-              {myUserId !== memberId && (
-                <>
-                  {reportStatus ? (
-                    <button
-                      onClick={() =>
-                        handleReportButtonClick(nickname, memberId)
-                      }
-                      className="btn btn-danger flex h-[1.875rem] w-[4.37rem] items-center rounded-lg p-0  px-[0.56rem] font-IMHyemin-bold text-sm"
-                    >
-                      신고하기
-                    </button>
-                  ) : rank < 500 ? (
-                    <span
-                      key={memberId}
-                      className="block h-[1.875rem] w-[4.37rem] text-center text-sm text-light-point dark:text-dark-point"
-                    >
-                      루틴 완료!
-                    </span>
-                  ) : isNotificationSent ? (
-                    <button
-                      key={memberId}
-                      className="btn btn-disabled h-[1.875rem] w-[4.37rem] cursor-default rounded-lg p-0 font-IMHyemin-bold text-sm"
-                    >
-                      내일 다시
-                    </button>
-                  ) : (
-                    <button
-                      key={memberId}
-                      className="btn dark:btn-dark-point btn-light-point flex h-[1.875rem] w-[4.37rem] items-center rounded-lg p-0  px-[0.56rem] font-IMHyemin-bold text-sm"
-                      onClick={() => handlePokeButtonClick(memberId, nickname)}
-                    >
-                      <Icon
-                        icon="BiSolidHandRight"
-                        size="lg"
-                        className="mr-[0.7rem]"
-                      />
-                      콕!
-                    </button>
-                  )}
-                </>
-              )}
-            </div>
-          )
-        )}
+        {members.map((member) => (
+          <RoomMember
+            member={member}
+            key={member.memberId}
+            reportStatus={reportStatus}
+            handleReportButtonClick={handleReportButtonClick}
+          />
+        ))}
       </div>
       <FormProvider {...form}>
         <ReportBottomSheet
